test: Add modified test for MPI_Win_test
While working on !25 (dedicated control messages for RMA synch), we
saw that an obvious bug in MPI_Win_test() was not caught by the test
suite. This commit adds the new test win_test_ordering.c for that bug.

This new test is derived from MPICH's rma/pscw_ordering.c but modified
so that it would run into a deadlock for the unfixed case of the bug.
(BTW: The bug didn't make it into the master.)