rma: Clean-up node_id handling for Win_allocate_shared()
This replaces the use of a hash on the node name with the use of that
id that is stored MPIDI_Process.smp_node_id (which is in the end the
node_id as determined by pscom).

In addition, the commit cleans up a bit the handling of shared-memory
communicators that have not been created via MPI_Comm_split_type().