cuda: Fix MPID_PSP_Reduce_for_cuda()
In the intercommunicator case, the root of a collective operation
is not determined by the rank within the local group but by the
special root value MPI_ROOT (cf. MPI-3.1, Sec. 5.2.3).
We have to take this possibility into account when handling the
receive buffer with regard to potential device memory. Otherwise,
tests like coll/icreduce may fail in a --with-cuda configuration.