rma: Cache information about node-local communicators
In MPID_PSP_Win_allocate_shared(), it must be checked whether the
given communicator is valid for being used with shared-memory RMA
windows, i.e. whether the communicator is host-local.

This information is now cached within the communicator after such
a check so that in a subsequent window creation with the same
communicator, this check can be omitted.