test: Add spawn_univ_discon.c to parastation tests
This test is the reproducer from Issue #9 that checks whether the
closing of connections may affect any received but still pending
Eager messages (i.e., pending in the unexpected message queue).

For this reason, it could make sense to vary/alter the Eager
threshold (e.g., via PSP_RENDEZVOUS) for multiple runs of the
test between the Eager and the Rendezvous range to check the
actual impact of unexpected messages on the test's behavior.