msa: Fix MPID_Get_node_id() for non COMM_SELF
In the case of topology awareness, we can use the badge table at
the nodes level. However, in the case without topology awareness,
we cannot provide valid information unless the ID of the own rank
is requested. Therefore, in all other cases, MPID_Get_node_id()
has to return a non-fatal error that can be handled in the higher
MPICH layers.