test/cuda: Rework laplace_mpi/simple_kernel tests
These tests didn't check for CUDA errors so that such could lead to
error messages supposedly related to the MPI layer. This commit
fixes this problem by explicitly checking for CUDA errors and
treating them as fatal if they would also affect the MPI layer.

In addition, the tests have been revised (also with respect to the
use of the mtest framework) and checked for algorithmic correctness.