init: Fix uninitialized value for singleton w/o PM
In the case of a singleton MPI process that was started without any
process manager, there is no Process Group (PG) it belongs to and
therefore also no pg_id_name. However, this string still needs to
be initialized properly and this commit does so by setting it to
an empty string.