hcoll: Fix zero-byte calls for inter-comm case
As HCOLL has problems with zero-byte messages, we use a shortcut in
such cases if HCOLL support has been configured (but not necessarily
activated via PSP_HCOLL=1 -- at least so far).

However, for the case of inter-communicators, we still have to handle
also zero-byte calls via a fallback to the pt2pt-based patterns as
provided by MPICH.

This commit fixes this issue by checking before doing the shortcut
whether the given communicator is HCOLL-enabled (which implies that
the communicator is an intra-communicator and HCOLL has been activated).
Otherwise, the shortcut is not used, but the fallback to the pt2pt2-
based patterns.