configure: Add 'gcc10' as new confset
The build system of MPICH 3.3.x fails to build with gcc version 10
since gfortran will by default here not compile files that call the
same routine with arguments of different types.

This commit adds the new confset 'gcc10' as a workaround for this
that implicitly just adds the "-fallow-argument-mismatch" compiler
flag in comparison to the common gcc/default confset.