test: Add test for Bcast with messages > 2 GiB
This adds a new test (large_msg_coll_bcast) that only checks for
the valid transfer of a message > 2 GiB via MPI_Bcast().

This test is to be run explicitly with one of the three algorithm
variants offered by MPICH for bcast (binomial, scatter_ring_allgather,
and scatter_recursive_doubling_allgather), see 'testlist' file.

Since such a bcast scenario was already present in 'large_msg_builtin'
(as one of several), this commit in addition switched to a point-to-
point scheme there to eliminate the influence of algorithm selection
at this point.