test: Add > 2 GiB test for non-contiguous datatypes
This adds a new test (large_msg_noncontig) that checks for the valid
transfer of a message > 2 GiB when using a non-contiguous datatype.
The test performs a broadcast-like pattern based on MPI_Send/Irecv
for checking pt2pt communication, an MPI_Bcast and MPI_Ibcast as
representatives for collective communication, as well as MPI_Put and
MPI_Get for checking also one-sided communication.

Since non-contiguous datatypes commonly require additional temporary
buffers, this test is marked as @largetest@ in the testlist file.