test: Rework TESTDIR/TESTSET feature
This reworks our implementation of the TESTDIR/TESTSET feature (cf.
make test) since there is now native support for choosing test lists
and test folders by the 'runtests' script.

Please note that this doesn't change the usage of TESTDIR/TESTSET
variables but allows in addition for comma-separated lists of folders
and test sets via TESTDIRS (note the 'S') and TESTLIST as the MPICH-
native variables for this.