test: Add TESTCUDA=no|yes option
By setting the new variable 'TESTCUDA' to 'no' when calling 'make test',
all CUDA features within the test suite can now explicitly be disabled
even if the MPI library has been built with CUDA awareness.

If this option is not set, the CUDA support of the MPI library is still
queried to decide whether CUDA-enabled tests should be used or not.