configure: Disable pscom4openib as "allin" default
Up to now 'pscom4openib' belonged to the list of pscom plugins that were
automatically being check for being built-in when '--with-pscom-allin'
is given without an explicit specification via '--with-pscom-builtin'.

However, since pscom4openib is now obsolete for production use, it is
also removed from this list of default allin plugins by this commit.