test: Add TESTCONF option.
With the new `TESTCONF` option, dedicated configuration sets can now be
selected to be used for test suite runs. Such a configuration is actually
just a file that is to be sourced before the call of the `runtests` script
and that may contain the set and export of certain environment variables
for certain purposes.

With this commit, the following test configuration sets are available:
*  hydra # Force the use of MPICH's Hydra mpiexec
*  psmgmt # Force the use of psmgmt's mpiexec in /opt/parastation/bin
*  verbose # Switch on verbose mode for 'runtests' script
*  psmgmt/memcheck # Call psmgmt's mpiexec with --memcheck
*  psmgmt/valgrind # Call psmpmt's mpiexec with --valgrind

Usage example: `make test TESTCONF=psmgmt/valgrind`