test: Replace mprobe.c by mprobe_threaded.c
The former `mprobe.c` in the parastation test folder was derived and
extended from an early version of `mprobe.c` in the pt2pt test folder,
and its extension was the use of threads. This commit reworks that
extension and removes those parts that were redundant in both tests.

Additionally, to make these changes more visible, the threaded mprobe
test in the parastation folder is now renamed to `mprobe_threaded.c`.