psp/hcoll: Remove query for PSP_HCOLL from PSP layer
Up to now, there was an explicit query in `MPID_Init()` for `PSP_HCOLL`
to enable HCOLL support and its use in the PSP layer.
However, as this may conflict with the settings of the MPICH environment
variables for that, this commit removes this explicit query within the
PSP layer so that HCOLL support now must be enabled also for the PSP
device via `MPIR_CVAR_ENABLE_HCOLL`.

In a subsequent commit, `PSP_HCOLL` is then to become an alias again for
`MPIR_CVAR_ENABLE_HCOLL`, but which is already evaluated by the upper
MPICH layer before `MPID_Init()` is called.