msa: Map MPID_Get_node_id() on MPIR_Process.node_map[]/MPID_Get_badge().
Since MPICH is now SMP-aware by default (comprising collective
communication), `MPID_Get_node_id()` can now directly make use of the
node map. This is created by the upper MPICH layers via PMI(x) and
stored in the `MPIR_Process.node_map[]` array -- at least if the
queried rank belongs to the local process group.

In the MSA case, `MPID_Get_badge()` is to be used instead  in the upper
layers with a fallback to MPICH's `node_map` if the badge table does not
exist at the node level.