msa/init: Do not fall back to APPNUM if no module ID is given.
Previously, we used the `APPNUM` to set the module ID if it was not set
by environment variable. Since this might lead to unexpected bahvior,
the following defaults are now applied for the module ID:

If MSA awareness is not enabeld, the module ID should still be `-1`.
Only if MSA is enabled but no module ID is given, it should be `0` as a
default.