vendor/hcoll: Switch from int to MPI_Aint for count arguments.
This commit adjusts the function signatures for the HCOLL collectives
from `int` to `MPI_Aint` for the `count` (and `displ`) arguments.

The HCOLL API still expects `int` for count arguments whereas MPID layer
uses `MPI_Aint` for this. Therefore, we should check whether we can
actually use HCOLL with a given count parameter.

In case of `Alltoallv()`, this even means to check all the array entries
and to agree on a global consensus via an additional `Allreduce()`, which
naturally impacts the latency.