finalize: Reactivate the possibility of a final barrier.
All previous versions of psmpi had a final barrier in `MPID_Finalize`.
With the transition to MPICH4, however, the world communicator is no
longer valid here, so that a barrier via `MPIR_Barrier` is no longer
applicable.

However, due to improvements in the shutdown protocol of pscom such a
barrier is also no longer necessary -- nevertheless might the former
behavior with a barrier at the end still be desired in certain cases.
This commit therefore adds a related callback to MPICH's finalize hook
infrastructure so that a barrier still can explicitly be activated.

The well-known `PSP_FINALIZE_BARRIER` environment variable can be used
to control this. Note, however, that the default values has changed to 0
(default = no barrier) and that the former (experimental) option to use
a "sparse" barrier has been removed.