psp: Enable re-init of psp device.
This commit enables the re-init of pscom connections by adapting
the exchange of connection information via the PM's global key-value
store (KVS). A unique key for the listen socket name per init is
created based on an init counter variable. In the finalize, a sanity
check in the cleanup of my_pg and a reset of next_lpid are added.

Remark: This solution needs to be revised for dynamic resources and
MPI sessions, especially for an expansion of the number of processes
over a re-init. In such a case, the counter variable will be different
in the original and new processes which will lead to incompatible keys.
This issue could be solved by using a per-process Publish/Lookup/
Unpublish mechanism instead of Put/Get on the global KVS.