vendor/hcoll: Call hcoll_finalize() when COMM_WORLD is being destroyed.
It is true that `hcoll_finalize()` needs to access `MPIR_Process.comm_world`
(via `rte_world_group_fn`) so that we can no longer call `hcoll_destroy()`
at the very end of the finalize procedure.

However, setting the priority in such a way that `hcoll_destroy()` is
called before the `MPI_COMM_WORLD` context is destroyed does also not
work out because then the related `hcoll_context_free()` call is no
longer valid.

The solution of this commit is hence to call `hcoll_destroy()` at that point
in time when `hcoll_comm_destroy()` has been called for `MPI_COMM_WORLD` --
assuming that this is also the last valid communicator context.

Please note that this doesn't seem to be an issue with vanilla MPICH
since there the deprecated `hcoll_destroy_context()` function is still
uses instead of `hcoll_context_free()`, which is apparently allowed to
be called after `hcoll_finalize()`.