msa: Fix mix-up between MSA_AWARENESS and MSA_AWARE_COLLOPS
Although both macros for encapsulating the MSA-related code parts
`MPID_PSP_MSA_AWARENESS` and `MPID_PSP_MSA_AWARE_COLLOPS` are set by
default if `--with-msa-awareness` is used, they do not necessarily mean
the same thing, since the first one is formally a superset of the other.

However, this proper distinction got somewhat neglected in the past,
and this commit seeks to fix this.