test: Add TESTCONFS option (list of test configurations).
Similar to the `TESTDIRS` option, this commit adds the option
`TESTCONFS` (note the "S" at the end) that allows the stating of a
comma-separated list of test configurations in comparison to the still
available `TESTCONF` option.

The difference between both, besides the possibility of using a list of
files, is that `TESCONF` still explicitly checks for the validity of the
given file, while `TESTCONFS` does not -- which is also quite similar to
the behavior of `TESTDIR` vs. `TESTDIRS`.