env: Move environment variable queries to MPID_Init().
This commit moves the querying of the following environment variables to
`MPID_Init()`:

`PSP_DEBUG`, `PSP_DEBUG_VERSION` and `PSP_HARD_ABORT`, as well as the
"hidden" `PSP_FINALIZE_*` options.

As a result, their querying now happens together with the other variables
in one central place. (However, further rework on this is to follow.)