test: Add TESTSETS option (list of test sets).
Similar to the `TESTDIRS` option, this commit adds the option `TESTSETS`
(note the "S" at the end) that allows the stating of a comma-separated
list of test sets in comparison to the still available `TESTSET` option.

The difference between both, besides the possibility of using a list of
test sets, is that `TESTSET` still explicitly checks for the validity of
the given set, while `TESTSETS` does not -- which is also quite similar
to the behavior of `TESTDIR` vs. `TESTDIRS`.