vendor/pmi: Fix pre-put KVS values for spawn_multiple
The pre-put KVS values must be specified for all sub-spawn commands of
a spanw_multiple, not just the first.

Apparently, Hydra has no problems with this, since the KVS seems to be
synchronized here, but with psmgmt, this information is missing for the
processes assigned to further executables.

This commit fixes this issue by removing the related case distinction
so that all processes always get this information.

The issue has been introduced with uptream commit 843f389.