psp/init: Remove counter variable from conn-info KVS keys.
In 37ff4138, a counter variable was added to create unique keys for the
exchange of connection information upon MPI re-init. This was meant as a
work-around for the restriction that overwriting existing keys with Put
operations on the KVS is not supported by all PM interfaces.

Since psmgmt supports duplicate keys for both PMI and PMIx and hydra
supports duplicate keys with MPICH PR #6564 (backported to psmpi in
previous commit), we can remove the counter variable from the key.

Remark: This solution can cause session re-init to fail for PM/ KVS
implementations that do not support duplicate keys.