psp/pg: Move init of global PG to MPID_Init.
This commit moves the initialization of the global PG MPIDI_Process.my_pg
from the MPIDI_PSP_Comm_commit_pre_hook of MPI_COMM_WORLD to a separate
function that is called in MPID_Init. Thereby, we ensure that the connection
information is stored globally in the PSP device independent of the init
of MPI_COMM_WORLD. This is a preparation to make communicator creation
independent of MPI_COMM_WORLD for MPI Sessions.