vendor/mpir_pmi: Fix re-init, connect to PM only once.
Upon re-init, we should not re-connect to the PM since the conenction
has not been teared down before (this happens in an atexit handler).
Not all PM interfaces behave in the same way if their init functions are
called multiple times from the same process. This solution mitigates
this issue by connecting only once to the PM per process - even if a
session re-init happens.