init: Refactor connection init.
Combine InitPscomConnections and InitPortConnections in one
function InitConnections with additional parameter ondemand.

This way, the redundancy of the code is reduced.

Place connection opening and connect in do_connect function to use this
in on-demand and non-on-demand cases.

Introduce do_connect_direct for non-on-demand case to call do_connect and
send init msg in addition.

Place exchange of connection data via KVS in to separate function.
Simplify and fix KVS key generation.