vendor/mpir_pmi: Support PMIx_Spawn in MPIR_pmi_spawn_multiple.
Add support for PMIx_Spawn in MPIR_pmi_spawn_multiple based on the
added utility functions.

The preput KVS values are stored in the KVS by the process calling
PMIx_Spawn using PMIx_Put prior to the call of PMIx_Spawn so that
the spawned processes can access the parent's preput values.

All error codes are set to the error code of PMIx_Spawn because
PMIx_Spawn does not provide error codes for individual apps.

In case PMIx_Spawn is not supported by the PMIx Host, a comprehensive
error string is returned to the user.