finalize: Disable finalize barrier for MPI Sessions.
For applications using only MPI Session (i.e., there is no
`MPI_COMM_WORLD`), the finalize barrier has to be disabled. This relies
on the existence of `MPI_COMM_WORLD`.

Similarly, the statistics feature (enabled by the
`--with-session-statistics` compile-time option) requires a global
`MPI_COMM_WORLD` communicator. Therefore, we have to ensure it exists in
the corresponding `print_stats` callback.