vendor/mpir_pmi: Fence with spawned processes for PMIx.
For spawning with PMIx we need to ensure that spawned processes
read the preput values from the KVS (i.e. the parent port string)
before the next spawn is launched. A PMIx_Fence with the processes
in the spawed namespace + the parent ensures that. The spawned
processes call PMIx_Fence after they read the preput value from
the KVS.