test: Allow also subdirectories for TESTDIR option.
Up to now, an error message was displayed if a (valid) subdirectory like
`threads/pt2pt` was specified with the `TESTDIR` option.
This commit fixes this issue and lets the `Makefile` also check for the
validity of test subdirectories.
In doing so, if only the given upper directory name is valid but not the
subdirectory, a list of all valid subdirectories within will be printed
together with the error message.