configure: Add RUNPATH for pscom to libmpi.so.
Up to now, `libmpi.so` does not have a `RUNPATH` attribute being set for
the respective path to the pscom, but only `mpicc` uses the `-rpath-link`
option to refer to it.

This commit changes this by evaluating the `PSCOM_RUNPATH` variable in
`psp/subconfigute.m4`. If not set, then the same path as used with `-L`
in the `PSCOM_LDFLAGS` is applied.

In addition, this commit removes the use of `-rpath-link` in `mpicc`, as
this has now become dispensable.