errors: Replace PRINTERROR macro with psp errnames.
This commit replaces the PRINTERROR macro with error reporting
based on the MPICH error stack. Errors specific for the psp device
are configured in mpich/src/mpid/psp/errnames.txt and can be used
to report errors occuring in the psp device.

There is more work required to add more psp-specific error messages
in the future.