test: Fix logfile name for TESTDIR with subfolders.
When the `TESTDIR` option (for specifying a single test folder) is
used, then this directory name becomes part of the logfile name.
However, when the given directory is actually composed of a folder
plus subfolder (which is possible since 857e15a8), then the logfile
name becomes invalid due to the splitting `/`.

This commit fixes this issue by substituting the `/` by a `-` for
the logfile name.