mpir_pmi: Do not provide PMIX_PREFIX to PMIx_Spawn.
We do not rely on OpenPMIx or the process manager to correctly
apply the path provided by a user on a spawn operation, but
instead we always add the path to the provided command and do
not forward the PMIX_PREFIX attribute to PMIx_Spawn.

This change was necessary because of different/ missing handling
of PMIX_PREFIX for different OpenPMIx versions and PMs.