vendor/mpir_pmix: Add PMIx-only spawn info keys for srun and mpiexec opts.
To support the setting of user-specific srun or mpiexec options for
spawned processes, this commit adds support for the experimental spawn
info keys srunopts and mpiexecopts to be used with ParaStation Management.

The new keys are only available with PMIx, not with PMI.

With these new keys, users can pass srun or mpiexec options to
MPI_Comm_spawn(_multiple) that will be appended to the srun or mpiexec
command by ParaStation Management.

Please take note of the comments on the limitations and risks of these
experimental keys that are added with this commit. Support for the
srunopts and mpiexecopts keys can be removed at any time in the future.

(Re-apply changes from 77801238.)