test: Add help for TESTDIR, TESTSET, and TESTCONF.
Up to now, when calling `make test` with `TESTDIR`, `TESTSET` and/or
`TESTCONF`, an overview of the available names is only given as part
of the error message when an invalid name has been stated.

This commit changes this by adding the possibility to use `help` as
a wildcard for the name to retrieve the overview lists without an
error message.

Examples:
> make test TESTSET=help
> make test TESTDIR=help
> make test TESTDIR=threads/help