test: Show explanation text in test conf list.
When using `make test TESTCONF=help` to retrieve the overview of
all available test configurations, now the first comment line in
the respective test configuration files is also printed behind the
config name.

For unifying this output a bit, this commit also modifies some of
the comments in the test configuration files.