vendor/errhan: Call errhandler of MPI_Session_init for invalid args.
In case of an invalid `info` or `session` argument, the autogenerated
`internal_Session_init()` does not evaluate the given `errhandler`,
but falls back to the default by leaving the associated pointer set
to `NULL`.

This commit fixes this by forwarding and evaluating the `errhandler`
explicitly (and again) in `MPIR_Err_return_session_init()` so that
it can be called there also in case of such invalid arguments.