init: Set MPI_UNIVERSE_SIZE via MPIEXEC_UNIVERSE_SIZE environment var.
The psp device provides the universe size based on what is provided
by the env variable MPIEXEC_UNIVERSE_SIZE or returns the value
MPIR_UNIVERSE_SIZE_NOT_AVAILABLE to the MPIR layer.

The MPI_UNIVERSE_SIZE attribute has the value of the env var
MPIEXEC_UNIVERSE_SIZE if this env var is set (either explicitly by
the user or via the process mananger). If the env var is not set,
the attribute MPI_UNIVERSE_SIZE is not available.

This solution circumvents requesting the universe size directly
from the process manager which may lead to different values
for PMI and PMIx, since for PMIx there is no attribute in the
standard that clearly matches the MPI_UNIVERSE_SIZE attribute.