test: Remove setting of MPIEXEC_UNIVERSE_SIZE in test suite.
Setting MPIEXEC_UNIVERSE_SIZE in the test suite is not required any longer.

Instead, the options -P -N 0 are passed to  psmgmt's mpiexec to reserve all
HW-threads on the current node (id 0). Together with PSI_OVERBOOK=1 (see
Makefile.in), this ensures sufficient resources for the spawn tests on a
single node.