init: Skip exchange of connection infos via KVS on re-init.
Store listen addresses of first init globally and add atexit handler
to free memory on exit. On re-init, these listen addresses are used
to connect to other processes and the exchange of listen addresses is
skipped.

This change is required for OpenPMIx 5 where overwriting of values
in the KVS for the same key is not supported. Furthermore, this
change is an optimization of the re-init procedure of psmpi.

Note that this commit removes pscom_stop_listen from the init procedure.
Hence, an application listens for incoming connections on the socket
for the lifetime of the application until the socket is closed which
happens either in pscom's atexit handler or during MPI_Finalize/
MPI_Session_finalize if PSP_FINALIZE_SHUTDOWN is set.