init: Introduce env var PSP_KEEP_CONNECTIONS.
If PSP_KEEP_CONNETIONS is set to >= 1, psmpi does not close the connections
to processes explicitly by calling pscom_close_connection(...) on
MPI_Finalize/ MPI_Session_finalize. Instead, the connections are closed in
the atexit handler of pscom where all remaining sockets and their connections
are cleaned up.

This is an experimental feature to optimize the MPI Session re-init where
connections can be re-used instead of being re-created.