init: Switch order of PG and topology init.
The topology needs to be initialized before the global PG to
have the topology information available for the global PG
init. This commit moves the topology init into MPIDI_PSP_PG_init
so that the correct order is ensured. Furthermore, the global
topology level list is removed from MPIDI_Process.