pg: Fix sorting in MPIDI_PSP_add_topo_level_to_pg().
The `MPIDI_PSP_add_topo_level_to_pg()` function shall add a new topology
level to the list associated with the respective PG in a sorted manner,
i.e. the new items are added in descending order with respect to their
degree value.

This commit fixes a bug that prevented this order from being maintained.