pg: Add more documentation to pg/topo_level functions.
This commit adds a few more explanatory words on the meaning and use
of the functions for managing the links between the process groups
(MPIDI_PG_t) and the topology levels (MPIDI_PSP_topo_level_t).

It also adds some comments to `MPIDI_PSP_add_topo_level_to_pg()`
that adds a new topology level to the list associated with a given
process group (pg).