connect: Fix singleton case, do not use KVS for 1 processs.
In case of only one process, there is no need to exchange
connection infos via the KVS.
Furthermore, MPIDI_Process.singleton_but_no_pm is removed
with this commit as it is a hack to derive the singleton
state from the appnum. Instead, the case of my_pg_size==1
is sufficient to skip any KVS activity.