init: Revise settings check, enable on default.
With this commit, the settings check makes use of the same
KVS barrier as the exchange of connection infos. Instead of
only comparing to the settings of rank 0, all processes
compare their settings to that of all other processes so
that all processes notice a settings mismatch and return an
error. The settings check is enabled on default.