vendor/info/mpix: Add MPIX_Info_merge_from_array() to API.
Up to now, the info objects of the MPI standard do not support the
concept of arrays. Of course, arrays can be encoded manually by
appending numerical values as indices to the keys. However, as
indexed values play a major role in the MPI world (think of process
ranks or the partitions of a buffer, for example), it may make sense
to also explicitly provide indexed values for info keys.

This proposal here adds `MPIX_Info_merge_from_array()` as a new
function to the API that can be used to create a single info object
from multiple info objects, where the values of identical keys are
_internally_ stored in an array.

Access to these arrays is currently only provided from _within_ the
MPI library (see `MPIR_Info_lookup_array()`), whereas on application
level, an access via the usual functions just reveals by the reserved
string `mpix_info_array_type` that the related key refers internally
to an array instead of a single value.