connect: Use pscom_resume_listen and pscom_suspend_listen.
The flag PSCOM_HAS_LISTEN_SUSPEND_RESUME is used at compile time to
check if the new pscom listen suspend/ resume API is available.

If it is not available a fallback to pscom_stop_listen/ pscom_listen
is used. This fallback is not safe and may lead to errors on MPI Session
re-init because the port assigned on first init via pscom_listen on
PSCOM_ANYPORT is returned to the OS with pscom_stop_listen and might be
used differently (busy) at re-init time.