make/test: Make mpiexec of test suite independent of build parameters.
Until now, the option chosen for the process manager (with or without hydra)
during the top-level configure was used to select the corresponding mpiexec
(`$(prefix)/bin/mpiexec.hydra` or `/opt/parastation/bin/mpiexec`).

With this commit, this magic is removed and the mpiexec that is in the `PATH`
or (if given) the one that is set via the `MPIEXEC` variable is selected.

If psmgmt or hydra are explicitly desired, the respective TESTCONF can still
be used.