make/test: Do some clean-up.
This commit removes the pointless use of `set`, some trailing whitespace,
and mutes the echo of the complex and quite longish command line used to
hand over to MPICH's make test, since the actual command to run the test
suite is echoed there anyway.
Additionally, error messages are now passed to stderr instead of stdout.