make/test: Add a check for a valid mpiexec.
This new check will warn the user if the test suite is run from the
top level Makefile with an mpiexec command (or similar) given via
`MPIEXEC` that cannot be found at this point.

However, we just warn and do not abort here, as the command may still
be set correctly in other ways for the lower-level scripts that run
the tests.