psp: Fix use of uninitialized value in MPIDI_PSP_Wait().
When `MPIDI_PSP_Wait()` is part of a call stack with a preceding
`MPIDU_Sched_progress()`, then the further invocation of the latter
within `MPIDI_PSP_Wait()` will directly return and `made_progress`
is not touched. However, since `made_progress` is not initialized
explicitly at the beginning of `MPIDI_PSP_Wait()`, its second `if`
statement can therefore depend on an uninitialized value.

This commit fixes this by setting `made_progress` explicitly to
zero at the beginning.