vendor/mpir_pmix: Re-apply PMIx logic for PMIx Process sets.
Re-apply the changes of 82d98e0a and adapt to the new file
structure. This includes the de-registration of event handlers in the
`atexit()` handler.

The compile-time checks for PMIx version 4 are removed because they
are no longer needed (PMIx version is already checked at configure
time, see c6b946b6d).