test: Make sure that Hydra's mpiexec is available when selected.
As Hydra is now disabled by default, the TESTCONF `hydra` (which by
its definition enforces the use of Hydra during the test suite run)
should make sure that the related mpiexec is actually available.
This commit adds such a check to the script for the `hydra` TESTCONF
so that it aborts with an error message and a hint to `make hydra`
if the check fails.