test/init: Remove tests exitst1, exitst2 and timeout from init/ps-test-minimal.
These tests focus on capabilities of the runtime (mpiexec, prterun) and not
on capabilities of the MPI library. Hence we remove them from the minimal
test set ps-test-minimal.

Furthermore, PRRTE (at least up to release 3.0.7) is not compatible
with runtest's resultTest=TestStatus feature and raises a false negative
test result for exitst1 and exitst2.