psp/rma: Relax constraints on win type for Win_shared_query().
Since MPI-4.1, the constraints on the type of a window when calling
`MPI_Win_shared_query()` have been relaxed and now allow it also to
be called for `MPI_WIN_FLAVOR_CREATE` and `MPI_WIN_FLAVOR_ALLOCATE`.

According to the standard, when a segment cannot be shared, "the
call returns size = 0 and a baseptr as if MPI_ALLOC_MEM was called
with size = 0" -- which in the end means `NULL` for `baseptr` and
that `disp_unit` isn't to be touched.

The latter also applies to the case of `MPI_PROC_NULL`, which is
why this commit also removes the setting of `*disp_unit = 0` for
this case.