configure: Enabled confset-independent customization of NVCC_FLAGS.
The customization of `nvcc` and the used flags (i.e., `NVCC_FLAGS`)
should be independent of the chosen confset. Now, if the user provides
customized `NVCC_FLAGS`, these are used for the test. If this fails, the
`configure` step will abort. If no `NVCC_FLAGS` are provided,
`configure` will try to fall back to `gcc` as host compiler.