test/parastation: Add MPI_Win_get/set_info() test for PSP case.
This commit adds the new test `win_info.c` to the `parastation/rma` test
folder that checks the desired behavior of `MPI_Win_get/set_info()` with
regard to setting (and also not setting) info keys for RMA windows.

Please note that this test is PSP/psmpi-specific and that the behavior
of MPICH differs here, so that the test would fail with CH3/CH4.