init: Make ondemand connections the default.
Ondemand connections are the default setting in production for
some time now. We switch to using ondemand connections by default
in psmpi with this commit.

In addition, the env var PSP_ONDEMAND is removed. It is replaced
by PSP_DIRECT_CONNECT for users to be able to explicitly activate
direct (immediate) connection establishment during MPI_Init_thread.

PSP_ONDEMAND_SPAWN is renamed to PSP_DIRECT_CONNECT_SPAWN with
inverted meaning.