configure: Use and check a variable with the ADI3 device name.
This commit is to clean up the top-level configure a bit.

The commit introduces a new variable dedicated to the device to be
used (psp as default, ch3/4 as alternatives) and checks at the end
for the SUMMARY output whether the PSP device was actually selected
despite any PSP options possibly still given. (Up to now, the output
in such cases may have been misleading.) If such options are selected
even though PSP is not being used, a warning is now issued, saying
that it is not certain whether these options (which are still passed
on to MPICH's configure) will have any effect.