vendor/ch4/init: Move hostname list generation into ch4 init.
The list of all hostnames is only required by the ch4 device to
enable shared memory for dynamic processes. Hence it can be moved
into the device layer so that other devices do not perform this
unnecessary MPI communication on startup.

Moreover, the procedure MPIR_nodeid_init caused a deadlock with
psmpi and --enable-msa-awareness when PSP_MSA_AWARENESS=0 (deactivated
at runtime). This deadlock is fixed with this commit.

Cherry-pick of MPICH commits:
- 1039d07: ch4: move MPIR_nodeid_init into MPID_InitCompleted
- 89a2ef4: mpir/nodemap: make MPIR_Process.node_hostnames optional