ucc: Add support for ucc in mpid/common.
This adds cross-device support for UCC to `mpid/common` alongside the
existing HCOLL support. The bindings design is broadly based on that
of HCOLL, but the implementation is also partly inspired by that found
in Open MPI.

Currently, only the following *blocking* collectives are supported:
 - MPI_Barrier
 - MPI_Bcast
 - MPI_Gather
 - MPI_Gatherv
 - MPI_Scatter
 - MPI_Scatterv
 - MPI_Reduce
 - MPI_Reduce_scatter
 - MPI_Reduce_scatter_block
 - MPI_Allgather
 - MPI_Allgatherv
 - MPI_Allreduce
 - MPI_Alltoall
 - MPI_Alltoallv

However, support for non-blocking operations should be quite straight-
forward to add.