ucc: Add support for ucc in mpid/ch4:ucx.
This adds the respective invocations of the UCC bindings to the UCX
netmod of the CH4 device. To activate this feature, `--with-ucc[=path]`
has to be used during configure and the environment variable
`MPIR_CVAR_CH4_UCC_ENABLE=1` has to be set at runtime.

Debug output and other messages from the bindings layer can be enabled
with `MPIR_CVAR_CH4_UCC_ENABLE_DEBUG` and `MPIR_CVAR_CH4_UCC_VERBOSITY_LEVEL`.