test/parastation/rma: Fix win_info test for multi-node case
Up to now, the multi-node case has not been taken into account in
the `win_info` test, but it was rather assumed that it is to be
executed only with a single process. Since `MPI_COMM_WORLD` was
generally used here, however, the test for `alloc_shared_noncontig`
led to an error in the multi-node case.

This commit fixes this by building a shared-memory communicator
first, which is then used for all sub-tests.

In addition, this commit fixes the number of processes to be used
with the `ps-test-minimal` test set to `1`.