configure: Add --enable-debug-mode option.
Up to now, it was only possible to enable the debug mode for the PSP
device and GCC via `--with-confset=debug`. This commit adds the new
option `--enable-debug-mode` which makes this mode (i.e., the configure
options for MPICH as specified in `CONF_OPS_DEBUG`) selectable also
for other confsets and devices such as CH3/4.

This commit also adds this option to the table in `README.md`.