spawn: Do not enforce using TCP on MPID_Open_port if RRComm is used.
RRComm precon and pscom's TCP plugin for payload communication are
mutual exclusive for now. This commit prevents that psmpi enforces
the use of the TCP plugin for spawning in case RRComm is used as precon.
We check for RRComm by checking if the process has stored an endpoint
string for itself at startup.

This commit moves the logic of enforcing the TCP plugin in certain cases
into a dedicated function and enhances the documentation.