configure/allin: Add rrcomm as builtin plugin for --with-pscom-builtin.
This commit enables the selection of the rrcomm precon interface of pscom
in an all-in build via --with-pscom-builtin=rrcomm.

First, the required rrcomm headers are searched in the system's include
path(s). If they are not found there, the default installation folder
/opt/parastation/include is checked.

This commit also fixes a bug with IOF marker in parsing the selected
built-in plugins and it prevents that a pscom rpath is added to the mpi
lib for pscom all-in builds.